pyproject.toml
[tool.poetry]
name = "spoffy"
version = "0.1.0"
description = "Spotify API client with async and sync support"
authors = ["Steinthor Palsson <steini90@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
urlobject = "^2"
attrs = "^19"
cattrs = "^0.9"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
flake8 = "^3.6"
mypy = "^0.660.0"
black = {version = "^18.3-alpha.4",allows-prereleases = true}
pytest-cov = "^2.6"
sphinx = "^2.0"
sphinx-autodoc-typehints = "^1.6"
sphinx_rtd_theme = "^0.4.3"
aiohttp = "^3.5"
requests = "^2.21"
astor = "^0.7.1"
pytest-asyncio = "^0.10.0"
snapshottest = "^0.5.0"
sphinxcontrib-asyncio = "^0.2.0"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
